Move VSTEffectOptionsDialog to a header file
